refactor: extract trace overlay lock command
